Implemented a Rudimentary SQLite3 DB Backend
I cannot stress enough that this is very rudimentary. It doesn't even have the ability to update the existing
database, if there is one. Your Database will have to be deleted each time you want to recreate it until such
time as I get around to implementing a sort of UPSERT, since SQLite3 does not support UPSERT natively.